🚧 Updated service for available characters module

+++ src/modules/client/available-characters/services/available-characters/available-characters.service.ts
@@ -1,10 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../../../../shared/prisma/prisma.service';
-import { AscensionMaterial, Prisma, PrismaPromise } from '@prisma/client';
-import {
-  IAvailableCharacter,
-  IAvailableCharacterCreate,
-} from '../../interfaces/common.interface';
+import { Prisma } from '@prisma/client';
+import { IAvailableCharacter } from '../../interfaces/common.interface';
 
 @Injectable()
 export class AvailableCharactersService {
@@ -22,30 +19,64 @@ export class AvailableCharactersService {
       .catch(() => null);
   }
 
-  public add(
+  public async update(
     userId: string,
-    characterIds: string[],
-  ): Promise<Prisma.BatchPayload> {
-    return this.prismaService.userCharacters.createMany({
-      data: characterIds.map((characterId: string): IAvailableCharacter => {
-        return {
-          userId,
-          characterId,
-        };
-      }),
-      skipDuplicates: true,
-    });
+    charactersToAdd: string[],
+    charactersToRemove: string[],
+  ): Promise<void> {
+    const operations: Promise<Prisma.BatchPayload>[] = [];
+
+    if (charactersToAdd.length > 0) {
+      operations.push(
+        this.prismaService.userCharacters.createMany({
+          data: charactersToAdd.map(
+            (characterId: string): IAvailableCharacter => ({
+              userId,
+              characterId,
+            }),
+          ),
+        }),
+      );
+    }
+
+    if (charactersToRemove.length > 0) {
+      operations.push(
+        this.prismaService.userCharacters.deleteMany({
+          where: {
+            userId,
+            characterId: { in: charactersToRemove },
+          },
+        }),
+      );
+    }
+
+    await Promise.all(operations);
   }
 
-  public async remove(
-    userId: string,
-    characterIds: string[],
-  ): Promise<Prisma.BatchPayload> {
-    return this.prismaService.userCharacters.deleteMany({
-      where: {
-        userId: userId,
-        characterId: { in: characterIds },
-      },
-    });
-  }
+  // public add(
+  //   userId: string,
+  //   characterIds: string[],
+  // ): Promise<Prisma.BatchPayload> {
+  //   return this.prismaService.userCharacters.createMany({
+  //     data: characterIds.map((characterId: string): IAvailableCharacter => {
+  //       return {
+  //         userId,
+  //         characterId,
+  //       };
+  //     }),
+  //     skipDuplicates: true,
+  //   });
+  // }
+  //
+  // public async remove(
+  //   userId: string,
+  //   characterIds: string[],
+  // ): Promise<Prisma.BatchPayload> {
+  //   return this.prismaService.userCharacters.deleteMany({
+  //     where: {
+  //       userId: userId,
+  //       characterId: { in: characterIds },
+  //     },
+  //   });
+  // }
 }
